outsourced optparse-applicative parsers to an own library module to access them in xerxes
